+++ src/ru/svetashov/geometry/ClosedLine.java
@@ -10,7 +10,7 @@ public class ClosedLine extends BrokenLine{
     public int length(){
         int res = super.length();
         if(points.size() < 3) return res;
-        return res + (int) Math.sqrt((points.get(points.size() - 1).x - points.get(0).x) * (points.get(points.size() - 1).x - points.get(0).x) + (points.get(points.size() - 1).y - points.get(0).y) * (points.get(points.size() - 1).y - points.get(0).y));
+        return res + (int) Math.sqrt((points.get(points.size() - 1).x.doubleValue() - points.get(0).x.doubleValue()) * (points.get(points.size() - 1).x.doubleValue() - points.get(0).x.doubleValue()) + (points.get(points.size() - 1).y.doubleValue() - points.get(0).y.doubleValue()) * (points.get(points.size() - 1).y.doubleValue() - points.get(0).y.doubleValue()));
     }
 
     public BrokenLine getBrokenLine(){
