+++ src/ru/svetashov/geometry/Square.java
@@ -35,9 +35,9 @@ public class Square extends Figure implements PolygonalChain{
     public BrokenLine getBrokenLine(){
         List<Point> points = new ArrayList<>();
         points.add(new Point(leftTop.x, leftTop.y));
-        points.add(new Point(leftTop.x + getWidth(), leftTop.y));
-        points.add(new Point(leftTop.x + getWidth(), leftTop.y - getWidth()));
-        points.add(new Point(leftTop.x, leftTop.y - getWidth()));
+        points.add(new Point(leftTop.x.doubleValue() + getWidth(), leftTop.y));
+        points.add(new Point(leftTop.x.doubleValue() + getWidth(), leftTop.y.doubleValue() - getWidth()));
+        points.add(new Point(leftTop.x, leftTop.y.doubleValue() - getWidth()));
         BrokenLine line = new BrokenLine(points);
         return line;
     }
